code/dte: update reference to the XML schemas of "factura electrónica"
Update the reference to the latest version of the XSD files
`DTE_v10.xsd` and `SiiTypes_v10.xsd`.
Both files are part of the set of XML schemas for 'factura electrónica'
and also for 'cesión (RTC)', there is a more recent version in the
second set of 'cesión (RTC)', registered in this repository.

Ref: https://github.com/cl-sii-extraoficial/archivos-oficiales/pull/14